+++ components/SocialMedia/index.ts
@@ -1,0 +1,3 @@
+import SocialMedia from './social-media';
+
+export default SocialMedia;
