+++ components/SocialMedia/parts/ModalSocialMedia/index.ts
@@ -1,0 +1,3 @@
+import ModalSocialMedia from './ModalSocialMedia';
+
+export default ModalSocialMedia;
