+++ components/SocialMedia/parts/ModalSocialMedia/styles.module.scss
@@ -46,7 +46,7 @@
 
 .members {
   @apply mb-5 text-base font-semibold underline text-center;
-  color: #9FA1AE;
+  color: #9fa1ae;
 }
 
 .description {
