--- components/SocialNetworking/index.ts
@@ -1,3 +1,0 @@
-import SocialNetworking from './social-networking';
-
-export default SocialNetworking;
