--- components/SocialNetworking/parts/ModalSocialNetworking/index.ts
@@ -1,3 +1,0 @@
-import ModalSocialNetworking from './ModalSocialNetworking';
-
-export default ModalSocialNetworking;
