+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
 import { usePreviewSubscription } from '@/lib/sanity';
 import { getClient, overlayDrafts } from '@/lib/sanity.server';
 import queries from '@/lib/queries';
-import SocialNetworking from '@/components/SocialNetworking';
+import SocialMedia from '@/components/SocialMedia';
 import DonationV2 from '@/components/DonationV2';
 import DonationV1 from '@/components/DonationV1';
 import Speakers from '@/components/Speakers';
@@ -80,7 +80,7 @@ export default function Index({
 
   return (
     <>
-      <SocialNetworking socialList={allSocialList[0].social_media_list} />
+      <SocialMedia socialList={allSocialList[0].social_media_list} />
       <DonationV2 donation={allDonationV2[0]} />
       <DonationV1 donation={allDonationV1[0]} />
       <Community community={allCommunity[0]} />
