+++ studio/schemas/schema.js
@@ -20,20 +20,20 @@ import merchGeneral from './merch-general';
 
 // Then we give our schema to the builder and provide the result to Sanity
 export default createSchema({
-    // We name our schema
-    name: 'default',
-    // Then proceed to concatenate our document type
-    // to the ones provided by any plugins that are installed
-    types: schemaTypes.concat([
-        /* Your types here! */
-        post,
-        author,
-        socialMedia,
-        donationV1,
-        donationV2,
-        speakers,
-        community,
-        partners,
-        merchGeneral,
-    ])
+  // We name our schema
+  name: 'default',
+  // Then proceed to concatenate our document type
+  // to the ones provided by any plugins that are installed
+  types: schemaTypes.concat([
+    /* Your types here! */
+    post,
+    author,
+    socialMedia,
+    donationV1,
+    donationV2,
+    speakers,
+    community,
+    partners,
+    merchGeneral
+  ])
 });
