+++ types/social-media.ts
